Don't bump version for @jupyterlab/rendermime-interfaces Fixes #11748
