+++ packages/rendermime-interfaces/package.json
@@ -30,7 +30,6 @@
     "watch": "tsc -b --watch"
   },
   "dependencies": {
-    "@jupyterlab/translation": "^4.0.0-alpha.9",
     "@lumino/coreutils": "^1.12.0",
     "@lumino/widgets": "^1.31.1"
   },
