+++ packages/translation/tsconfig.json
@@ -10,6 +10,9 @@
       "path": "../coreutils"
     },
     {
+      "path": "../rendermime-interfaces"
+    },
+    {
       "path": "../services"
     },
     {
